chore(PR): Move RSA pub/private keys to a constants file
**Tickets:**
* [sc-82477](https://app.shortcut.com/publicsq/story/82477)
